fix: Rework the FulaOrdAPI to be more stable
Since to API took a string as an parameter and filtered the bad words out of it it sometimes didn't work and crashed. Now the the application only gets the bad words from the API and uses that list to sort out bad words itself. Now the filtering is a lot more stable.